[CSAPI] consolidate all files into one
